requirements.txt
pyro-ppl~=1.8.1
torch~=1.12.0
matplotlib~=3.5.2
pandas~=1.4.3
numpy~=1.23.1
tqdm~=4.64.0